[WIP] Read other transforms and print Bones to .cae
This change allows the program to read the other 2 sequences of 4x4 matrices that follow the initial Bone definitions. These matrices might be important for the armature working properly, but I don't yet know how any of them fit in together.
Also, refactor the 4x4 matrix into its own class to make it easier to use.